`mypy_test.py`: Always add dependencies of stubs to the files to test
