+++ tests/regr_test.py
@@ -9,7 +9,7 @@ import shutil
 import subprocess
 import sys
 import tempfile
-from itertools import filterfalse, product
+from itertools import product
 from pathlib import Path
 from typing_extensions import TypeAlias
 
@@ -17,9 +17,9 @@ from utils import (
     PackageInfo,
     colored,
     get_all_testcase_directories,
+    get_recursive_requirements,
     print_error,
     print_success_msg,
-    read_dependencies,
     testcase_dir_from_package_name,
 )
 
@@ -79,13 +79,6 @@ parser.add_argument(
         "Note that this cannot be specified if --all is also specified."
     ),
 )
-
-
-def get_recursive_requirements(package_name: str, seen: set[str] | None = None) -> list[str]:
-    seen = seen if seen is not None else {package_name}
-    for dependency in filterfalse(seen.__contains__, read_dependencies(package_name)):
-        seen.update(get_recursive_requirements(dependency, seen))
-    return sorted(seen | {package_name})
 
 
 def test_testcase_directory(package: PackageInfo, version: str, platform: str) -> ReturnCode:
